+++ twitchAPI/twitch.py
@@ -3558,8 +3558,8 @@ class Twitch:
         For detailed documentation, see here: https://dev.twitch.tv/docs/api/reference#get-soundtrack-playlist
 
         :param broadcaster_id: The ID of the playlist to get.
-        :param after: The maximum number of items to return per page in the response. Between 1 and 50. |default| :code:`20`
-        :param first: The cursor used to get the next page of results.
+        :param first: The maximum number of items to return per page in the response. Between 1 and 50. |default| :code:`20`
+        :param after: The cursor used to get the next page of results.
         :raises ~twitchAPI.types.TwitchAPIException: if the request was malformed
         :raises ~twitchAPI.types.UnauthorizedException: if user authentication is not set or invalid
         :raises ~twitchAPI.types.TwitchAuthorizationException: if the used authentication token became invalid and a re authentication failed
@@ -3575,4 +3575,33 @@ class Twitch:
             'after': after
         }
         async for y in self._build_generator('GET', 'soundtrack/playlist', param, AuthType.EITHER, [], Soundtrack):
+            yield y
+
+    async def get_soundtrack_playlists(self,
+                                       playlist_id: Optional[str] = None,
+                                       first: Optional[int] = None,
+                                       after: Optional[str] = None) -> AsyncGenerator[Playlist, None]:
+        """Gets a list of Soundtrack playlists.
+
+        Requires User or App Authentication\n
+        For detailed documentation, see here: https://dev.twitch.tv/docs/api/reference#get-soundtrack-playlists
+
+        :param playlist_id: The ID of the playlist to get. Specify an ID only if you want to get a single playlist instead of all playlists.
+        :param first: The maximum number of items to return per page in the response. Between 1 and 50. |default| :code:`20`
+        :param after: The cursor used to get the next page of results.
+        :raises ~twitchAPI.types.TwitchAPIException: if the request was malformed
+        :raises ~twitchAPI.types.UnauthorizedException: if user authentication is not set or invalid
+        :raises ~twitchAPI.types.TwitchAuthorizationException: if the used authentication token became invalid and a re authentication failed
+        :raises ~twitchAPI.types.TwitchBackendException: if the Twitch API itself runs into problems
+        :raises ~twitchAPI.types.TwitchAPIException: if a Query Parameter is missing or invalid
+        :raises ValueError: if first is not between 1 and 50
+        """
+        if first is not None and first < 1 or first > 50:
+            raise ValueError('first must be between 1 and 50')
+        param = {
+            'id': playlist_id,
+            'first': first,
+            'after': after
+        }
+        async for y in self._build_generator('GET', 'soundtrack/playlists', param, AuthType.EITHER, [], Playlist):
             yield y